requirements.txt
fastapi[standard]
uvicorn
requests
pydantic_settings
streamlit
langchain
langchain-openai
langchain-community
sqlalchemy
psycopg2-binary
pyjwt
passlib[bcrypt]